refactor repl
